Security fix: Paths were not sanitized on Windows. Fixes #314
Security related: Added sanitizing of paths to avoid false positives on
coderisk.com (there where no risk because already test the paths for
sanity - but this is not detected by coderisk, as the variable is not
modified). This was simply done in order get rid of the warnings at
coderisk. Fixes #312